transform component, scene testing
* removed position,scale,rotation components
* added Transform component that contains position, rotation, and scale
* scene graph update now regenerates the local mat4 using the transform
component
* akara bugfix: adding new subscriptions will process existing entities
* added `--testscene` arg for testing individual scenes in isolation
* added rotation support to d2interface.Surface

+++ d2common/d2sprite/sprite.go
@@ -83,7 +83,7 @@ func (a *Sprite) Advance(elapsed time.Duration) error {
 	a.lastFrameTime += elapsed
 
 	framesAdvanced := int(float64(a.lastFrameTime) / float64(frameLength))
-	a.lastFrameTime -= time.Duration(float64(framesAdvanced) * float64(frameLength))
+	a.lastFrameTime -= time.Duration(float64(framesAdvanced)) * frameLength
 
 	for i := 0; i < framesAdvanced; i++ {
 		startIndex := 0
